Moved out ItemEnumFunctions
