+++ src/Engine/Objects/Items.h
@@ -4,6 +4,7 @@
 #include <optional>
 
 #include "Engine/Objects/ItemEnums.h"
+#include "Engine/Objects/ItemEnumFunctions.h"
 #include "Engine/Objects/CharacterEnums.h"
 #include "Engine/Objects/SpriteObjectEnums.h"
 #include "Engine/Time.h"
